Add optional handling of meta files

+++ src/commands/createClass.ts
@@ -85,8 +85,13 @@ export default function createClass() {
     }
 
     function generateFile(classname) {
-        //return Promise.all([writeFile(), writeMetaFile()]);
-        return Promise.all([writeFile()]);
+		
+		if (vscode.window.forceCode.config.handleMetaFiles) {
+			return Promise.all([writeFile(), writeMetaFile()]);
+		} else {
+			return Promise.all([writeFile()]);
+		}
+		
         function writeFile() {
             return new Promise(function (resolve, reject) {
                 // Write Class file
@@ -114,42 +119,36 @@ export default function createClass() {
                 });
             });
         }
-        /*
+
         // Write Metadata file
         function writeMetaFile() {
             var finalMetadataName: string = classesPath + path.sep + classname + '.cls-meta.xml';
             return new Promise(function (resolve, reject) {
                 fs.stat(finalMetadataName, function (err, stats) {
                     if (!err) {
-                        vscode.window.forceCode.statusBarItem.text = 'ForceCode: Error creating file';
                         vscode.window.showErrorMessage('Cannot create ' + finalMetadataName + '. A file with that name already exists!');
                     } else if (err.code === 'ENOENT') {
 
                         var metaFile: string = `<?xml version="1.0" encoding="UTF-8"?>
 <ApexClass xmlns="http://soap.sforce.com/2006/04/metadata">
-    <apiVersion>${config.apiVersion || vscode.window.forceCode.conn.version || '37.0'}</apiVersion>
+    <apiVersion>${config.apiVersion || vscode.window.forceCode.conn.version || '43.0'}</apiVersion>
     <status>Active</status>
 </ApexClass>`;
 
                         fs.outputFile(finalMetadataName, metaFile, function (writeError) {
                             if (writeError) {
-                                vscode.window.forceCode.statusBarItem.text = 'ForceCode: ' + writeError.message;
                                 vscode.window.showErrorMessage(writeError.message);
                                 reject(err);
                             }
                             resolve(finalMetadataName);
                         });
                     } else {
-                        vscode.window.forceCode.statusBarItem.text = 'ForceCode: ' + err.code;
                         reject(err);
                     }
-                    vscode.window.forceCode.resetMenu();
                 });
 
             });
         }
-        */
-
 
     }
